services/follow-api/requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
pydantic
pika
python-jose